add some helper methods for writing features compactly
add some simple helper methods to store common data types (strings,
unsigned integers) in compact, variable-length encodings to keep
stored features as compact as possible

many of the string features we stores are relatively small (e.g.,
a single token), as are all of the identifiers we store (e.g., enum
IDs); however, I don't want to hard-code the serialization format
with these assumptions because we'll probably violate them at some
point in the future.

instead, add a simple MSB-extended 7/1 unsigned encoding, so that
small values take a small amount of space, and large values take
more.

 * 1B for values 0 .. 127
 * 2B for values 128 .. 16Ki - 1
 * 3B for values 16Ki .. 2Mi - 1
 * 4B for values 2Mi .. 256Mi - 1
 * 5B for values 256Mi .. 4Gi - 1

this can be trivially extended to support larger values, too, if
we encounter any situations where we would like to be able to
store fields that may need long values but the typical value is
small.