ml-common: add helpers to validate reflective method bindings
the feature pipeline allows dynamically binding the feature types
produces by multiple pipeline stages as the input arguments for
dependent pipeline stages.

if the pipeline is mis-specified, this will result in MethodInvocationErrors
(or other run-time exceptions) that will be extremely hard to triage and
debug

to help avoid these errors, add some helper methods to support validating
if a list of Types (ie, the bound inputs to a pipeline stage) conform to
the type signature of a method (ie, the specific transform's apply method).
getting this to work required diving deep into the depths of Scala's
internal (and fairly poorly documented) Type hierarchy

various links that were helpful in developing these methods:

* http://grokbase.com/t/gg/scala-user/151qje1xvx/how-to-detect-convert-t-type-to-seq-t-in-a-macro
* http://www.scala-lang.org/api/current/scala-reflect/index.html#scala.reflect.api.Universe

these methods will be used by FeaturePipeline to validate that the
connections specified in the pipeline JSON result in a valid
pipeline object at load time