ml-feature: bind the feature pipeline into a callable graph
applying the feature pipeline to a document entails calling the
transformations within each pipeline stage in turn, copying
their results into an intermediate data structure, then
selecting the items in the itermediate data structure that are
needed as input for each transform in the next stage and calling
those transforms. this repeats until the pipeline completes

to avoid a lot of repeated reflective look-up operations at apply-
time, the feature pipeline "binds" the FeatureTransform graph by
generating thunk functions for each transformer which know how
to deserialize the intermediate data into appropriate input
arguments for the reflective call.

furthermore, since pipeline graphs may become quite large and
contain lots of intermediate results, the bind process also computes
the "dead" intermediate values at each transform stage (i.e., the
inputs computed in previous stages that are not referenced as inputs
by any transforms in future stages), and stores this list with
the bound stage to prune the intermediates table

graphically, given a feature pipeline:

[{ name: 'A', inputs: [ '$document' ] },
 { name: 'B', inputs: [ 'A' ] },
 { name: 'C', inputs: [ 'B' ] },
 { name: 'D', inputs: [ 'A' ] },
 { name: 'E', inputs: [ 'C', 'D' ] }]

i.e.:

      B  -  C
    /         \
  A            E
    \         /
      D -----

the feature transforms will be sorted into independent pipeline
stages as:

  [ [ 'A' ],
    [ 'B', 'D' ],
    [ 'C' ],
    [ 'E' ] ]

the binding generates thunks for each transform in each stage which
calls the apply() method for each of the transforms in that stage with
the correct inputs for each transform, i.e.

  [ [ thunk:A.apply(document) ],
    [ thunk:B.apply(A.output), thunk:D.apply(A.output) ],
    [ thunk:C.apply(B.output) ],
    [ thunk:E.apply(C.output, D.output) ] ]