ENG-2916: unit-testing support for log messages
add a new helper trait that can be mixed in to unit test classes
to verify that log messages are being created as expected

to use in a test class, mixin the BeforeAndAfterAll and
BeforeAndAfter traits from scalatest, override loggername, and
initialize / shutdown the injected logger in the beforeAll and
afterAll helpers (respectively), e.g.:

  class MyFeatureSpec extends... with VerifyLogging
    with BeforeAndAfterAll with BeforeAndAfter {

    override val loggerName = "com.idibon.ml.feature.MyFeature"

    override def beforeAll {
      super.beforeAll
      initializeLogging
    }

    override def afterAll {
      shutdownLogging
      super.afterAll
    }

    after {
      resetLog
    }
  }

then, throughout the test suite, you can use the loggedMessages
property on the trait to look at what messages have been logged
since the last time resetLog was called.

also, clean-up the project dependencies to ensure that slf4j layers
on top of log4j 2.5, rather than 1.2.