idiml-cld2: fix CLD2 Linux g++ build
the CLD2 wasn't spec-compliant C++, since it used the "FILE"
type without actually including the header file that declared it.

after fixing this issue (and changing the stdint.h include to
the correct C++ header, cstdint), g++ complained that the headers
require compilation in C++11-compliant mode, so I added this flag
to the compiler arguments

and, after adding the C++11 flag, the compiler complained that the
auto- generated UTF-8 property table was using narrowing conversions
(specifically using negative values in an initializer list for an
unsigned table). this is illegal in C++11 (and was undefined behavior
in earlier versions), so I used regexp-replace to convert all of the
negative values to their correct 2s complement versions.

after tracing through the code, it looks like the negative offset
portions of this table are accessed through a reinterpret_cast<int8*>,
so storing the 2s complement unsigned value in the table is correct

with these changes, CLD2 compiles on linux and OSX, and works with
manual english, french, chinese, japanese, danish, hebrew, russian
and tamil tests inside scalaConsole