Add almost working DocumentRules
Specifically this:
 - fleshes out the DocumentRules class
 - adds SafeCharSequence Class
 - and uses Java Pattern more explicitly
 - adds unit test stubs

This change more closely models the "document" rule path that we
have in the JRuby code. It does not do the complex things that are
required for spans, though. It uses the scala Try concept to handle
cases where the regular expression compilation barfs or we timeout.
TODO:
 - It requires unit tests
 - being able to save
 - and load