Adds save, load and unit tests to DocumentRule model
Saving and loading a rule model should now be possible. Unit tests
are also added. This also addresses some comments from the previous commit.

Saving is based on turning the rules into JSON and saving that
to a resource named rules.json. The label is saved as config. This
should make it easy to inspect the file, and then hydrate a rule model
from a save rules.json file.