Adds fleshed out Ensemble model
This change fleshes out the ensemble model.The ensemble model in spirit
is just delegation, so should be fairly straightforward. However
it needs some "combiner" logic.

So a weighted average combiner that mirrors our current JRuby
implementation is added. It special cases the white or black list rule
override, which is nasty. I'm sure there could be a better abstraction
to handle it... Otherwise it does a weighted average as described
here https://idibon.atlassian.net/wiki/display/DSci/Model+Training+Process#ModelTrainingProcess-Document-ScopeTasks
point 6. Note: this is a weighted average between all models, which should be
two for most cases, but just mentioning that this will do weighted average
between N models, which in our world should be just two: LR model and Rule model.

Also I have extended IntentAlloy a little so that it can run tests in parallel
without running over itself, mainly by prepending a random number to the name
of a file so that individual instantiations write to different files...

This also refactors the returned "result" object from DocumentPredictionResult
to PredictResult and SingleLableDocumentPredictResult. I was essentially making
a case class, so I converted it to one with a hierarchy, while making it simpler
by only handling a result for a single label. Pushing the top level abstract PredictResult
as the general top level return type might be a little aggressive, but would allow for
future case class business for multi-label results...