idiml-features: port all of the Feature code over to Built
remove save and load from the Feature interface, and migrate all existing
classes to be defined as "class A extends Feature[X] with Built[A, B]", where
B is a Builder for A.

in most cases, Transformer code just process features in-flight, and doesn't know
or care about the Built trait, so no changes are needed. however, one special case
is introduced in IndexTransformer where it is now possible that some of the input
Feature instances do not also implement Built, thereby rendering the features
un-saveable.

this also makes all of the existing Feature classes just stores for
immutable data, so make this change, too