Refactor predict method to take PredictOptions object
As Gary pointed out, so the method signature doesn't get too crazy,
add a class that is basically a map that contains these prediction parameters.

So I made this class wrap a map, that takes enum values as keys, with an
Any type in the value. So clients of this must do their own casting
of values from the options to make use of it.