Refactor PredictOptions to be simpler case class
Since the enum and map were overkill, simplify this object a bit.
Also instead of two things for significant features, just use the
significantFeatureThreshold value to indicate whether they should
be returned or not. The "off" value is Float.NaN and to check
for it you need to use the .isNaN() function. This means that
significant feature values can be any Float value, while making
it easy to know whether to return them or not.