ENG-2987: separate Archivable#load into ArchiveLoader
migrate FeatureTransformer and Model classes to use the paired-builder
trait pattern, so that more of our classes can be defined as immutable
objects

this also updates every class except IndexTransformer to use
immutable collections internally; IndexTransformer can not use immutable
maps (at least in its normal instantiation) because the map is built as
the transform is applied.

this also modified the PredictModel trait to not extend from Archivable,
to prevent continuous repetition of the type parameters. each model is
responsible for extending Archivable directly (which means that some
models may not be Archivable)

I ended up doing a bit more refactoring here than I would like in a
single commit -- this cleans up (and actually parallelizes) some of
the precompilation for DocumentRules and moves some shared-ish
instantiation code for transform and model loading into the
Archivable and ArchiveLoader companion objects