ENG-2988 Add code coverage tool to idiml project
Code coverage is not configured at the build.gradle top level, but in each subproject,
since that's the only way for it to work.
We chose to use https://github.com/scoverage/gradle-scoverage because it was the first
thing suggested and I managed to get it working quickly. Otherwise no other options
were considered. There is optional configuration for not making builds pass, if
a certain coverage isn't obtained, but I didn't look into that.

This adds a few gradle targets that can now be invoked. To get it to run I did:
 > ./graldew reportScoverage
which produces the HTML report. Since this also runs unit tests, I am overriding
the circle.yml config to use it to run the unit tests.