Removes the java src dirs and moves any java code to scala src dirs
Since we're referencing java code from scala and vice-versa, to compile
that we need to enable join compilation. Joint compilation only works
if you have everything under the same source directory, or you tell
gradle that the java source is also a scala source set, and blank out the
java source set.

So instead of doing gradle configuration, we opted to just merged
the source trees under the scala source. This has the added benefit
of everything in the same package living in the same directory.