Adds ScalaJarAlloy and removes JarAlloy
This change does a bit of refactoring, settling on a scala version
of the JarAlloy, since that would require less type manipulation/refactoring
of existing methods to help reify a model.

So currently ScalaJarAlloy correctly saves and loads models. Otherwise
it still is WIP in terms of making it contain everything we require
for making Task based predictions.