Refactors train Engine to output MVP idiml alloy jar
With this change we can now take data from production,
e.g. `./export_training_to_idiml.rb -e production -k $IDIBON_PROD_KEY
-c ESS_Reaggregated -t Sentiment -o ~/temp/ess_idiml.txt -a simple`
and use it as input for training a model and getting an alloy jar
back out.

So this refactors a bunch of files for that to work, and also does
some minor clean-up of IntentAlloy (moved to test) and renaming
IndexerSpec to fit with the standard <class nam>Spec convention.
It also tweaks unit tests appropriately.

Decisions to note:
 - the parameters here for training the model are garbage. We need
  to decide how to programmatically set them and what are good defaults.
 - feature pipeline is hardcoded for training.
 - we need to check training sets are created as intended.
 - need to know how to handle multiple positive annotations (e.g. fix
  the data generation script or handle it here).
 - crude OOV handling: if there are OOV words, the feature index just
  passes them through. To handle prediction not breaking, I decided to
  cut the 'new' dimensions off and create a new feature vector. Perhaps
  the featureIndex should take care of this instead (no unit tests on
  this bit BTW).