Address review comments and refactors Engine traits
I refactored the engine traits so that training and prediction were
different traits. Those engines however extend from the same BaseEngine
which extends from the original Engine trait and controls the single
instantiation of the spark context. This way the engines are decoupled,
but the instantiation of the spark context is in a single place.

Othwerise the rest of this commit mainly adds some better logging and
changes the predict engine to take in a file to predict over.