Refactor pipeline apply to return single vector
Add "prime" function to warm up the feature pipeline. This
then stores the requisite information about dimensions of
the transformers. Adds function numDimensions to a FT trait,
so only those feature transforms that need to, will implement
this.

Change the behavior so that a single feature vector is output
from the feature pipeline. It internally stores the offsets
and dimension lengths of the vectors for the output feature
transformers, so we can calculate the correct indexes.

Note: this ASSUMES that the order given by the PipelineEntry
for the $output stage, is the same as the order in which the
vector sequence is output and thus concatenated.

This change is also backwards compatible with anything that's
existing, except that before using a feature pipeline prime
must be called, or if loaded from a jar, it is assumed it is
already primed. If not "prime" can be run again to update the
values.

Next stop: "purging" what's stored in the feature transformers.