improve gradle use in CircleCI
add a new task that actually downloads and resolves dependencies,
so that we don't spend 2 minutes on every change downloading
static dependencies

add "--console plain" to all of the Gradle command lines so that
the console capture in CircleCI doesn't get confused by all of
the formatting characters (backspace, etc) that Gradle uses in
the default "rich" mode. the huge log spew in "rich" mode on
Circle kills Firefox, and plain mode does not.