Optimizes getFeatureVector() in IndexTransformer
Previously IndexTransformer would create a dense array of the
entire vocab size. i.e. if we had 1 million words, we would
try to allocate an array of size 1 million. Which was all for the
purpose of providing up to n, feature index values, where n << |V|
in practise. This change makes it so we construct the appropriate
arrays ourselves for creating the sparse vector. No unit tests
need to be added since there was already good coverage here.