Address PR comments
The WordShapesTransformer functionality has been
changed to ignore counts of unique shapes, now returning
a sequence of shapes (no longer the unique shapes) in
the order they are appear. As a result, the Shape class has
also been updated to not include counts.

The regexes used to convert a word into it's shape string
equivalent have all been updated to support unicode.