Adds significant features to LR model
This makes the LR model wrapper compute the significant features.
It then wires through the ability via FeatureTransfomer to go
back from an integer feature dimension index to a human readable
string. Unit test coverage is also added.

I also started the convention that a feature that is expected
to be presented to humans must override the toString method
and prefix the value with the type of feature.