Adds rules to training a model via the training app
If you want to package rules up in an alloy while training
a model, you can now do that by passing in a JSON file. The
file is the output of the export training data to idiml ruby
script. But in short it's a single rule per line, where each
line is a JSON object with fields label, expression and weight.

Unit tests forth coming (well more like integration tests).