Refactor IndexTransformer and FeaturePipeline and FeatureTrait
This change makes build() produce an unfrozen feature pipeline
while prime() & thus load() produce frozen pipelines. This is
so that vector concatenation can be done easily (note: I need
to fix the unit tests for this).

I introduce the Terminable Feature trait to better segregate that,
with the addition of a "freeze()" method. That then is implemented
in the IndexTransformer.

So I refactored IndexTransformer to now know when it is frozen,
and when it can or cannot add to itself.

I am probably missing a unit test or two, but otherwise things
run without complaint.