ENG-3071 Make frozen idempotent in indexTransformer
The bug was that on loading a feature pipeline, we call freeze()
again, which calls freeze on the indextransformer. The freeze
in the indexTransformer was not idempotent, and was resetting
the "dimensions" to the current index size. This change makes
it idempotent and adds integration test coverage around that.

I also fixed a bug around not handling an entire vector of
OOV terms. Added unit test for that.