Refactors the trainer and casts furnaces
Furnaces are what create the MLModels required as part of an
alloy. This refactor introduces a bunch of new traits and classes.
It also takes https://github.com/idibon/idiml/pull/60/files and
incorporates it here - i.e. L1 defaults, and saving training data
locally to parquet.

The aim of this commit is to separate training concerns so that we can
extend modules easily and without much effort.

AlloyTrainer - is a trait that has top level signature needed to be
  implemented, as well as implementation agnostic functions.

BaseTrainer - outlines a fairly orthodox training algorithm that classes
  extend and fill out. E.g. KClass1FP trains K models with 1 feature pipeline.

Furnace - trait that different MLModel making strategies extend and fill in.
E.g. XValLogisticRegression which extends LogisticRegressionFurnace.

Also refators RDDGeneration into a trait that can be extended by other
strategies for creating RDDs.

Things left to be done:
 - passing through feature pipeline configuration
 - dealing with configuring the Furnaces, e.g. xval params.