Refactors RDDGenerator to be more generic
This moves the parquet creation into the former RDDGenerator which
is now SparkDataGenerator[T]. It being parameterized means that
we could support MLLIB a bit more "natively" instead of converting
from DataFrames - but I might be over-engineering and it wont be an
issue. But for now everything outputs dataframes instead of RDDs.

I have a FIXME to write unit tests for the new spark generator.