Adds MLLIB LR Extension class
This is a psuedo wrapper, since we extend the Spark MLLIB LR
class to expose prediction on a single vector that returns
probabilities for all classes. Adds unit test coverage as well.

Note: we could probably heavily change the way the internal
representation works, since this model relies on dense vectors.
Dense vectors are memory hogs, so re-writing prediction to work
on sparse vectors might be a good idea to do in the future. Also
the code to compute could be simplified. See hidden comments in
https://github.com/idibon/idiml/pull/63/files.