Adds IdibonMultiClassLRModel
This adds the IdibonMultiClassLRModel that wraps the MLLIB multinomial
LR wrapper class and makes it useable. This change also adds the ability
to get the correct features used by the MLLIB model, since that was not
exposed before, and we need to manually do a few things to recover
what feature indexes are actually used by the MLLIB multinomial LR model.

Comes with unit test coverage.