Adds GangModel
This is a model that wraps a multiclass model and per label models.
e.g a multinomial LR model + per label rule models. It is envisaged
that there is only ONE GangModel per task.

This model requires that the multiclass model is provided, otherwise
the per label models are not needed. It is also under the assumption
that the multiclass model returns a MultiLabelDocumentResult.

This also adds to the BaseAlloy the ability to then get predictions
from all models loaded by an alloy. I.e. if there's only a single
gangmodel all label results are returned, else each per label
model is called; this makes use of the earlier PredictResult additions
to take care of this.