Adds test app for experimentation
Until we have buidlers setup to be able to configure what we want
to build and predict on, I wrote this small app to allow myself to
do that. It doesn't save the alloy, it just trains, creates it and
runs a prediction against it. It helped me to iterate and make
sure end to end pieces worked without having to write the save &
load pieces... I expect this to disappear or become some experiemental
entry point into our code...