Adds furnace builders
Adds a trait that the furnaces extend from, as well as LR
specific traits. Then I add three classes, one for each
LR furnace type that we have. I also setup a static object
to house the common LR defaults.

The neat thing done here is that we can instantiate these furnace
builders from JSON config - and to not deal with reflection foo
I am telling the JSON parser what type of class it could be. I
think this should be fine to maintain, as you'll only need to
add it once.

The next idea is to do something simliar for the DataGenerators,
and the AlloyTrainers.