Refactors ArchiveLoader Trait to take an Option[Alloy.Reader]
This is so that we can use the same FeaturePipeline JSON config
structure for creating a BRAND NEW pipeline. Since the assumption
for the loaders was that they are loading from an Alloy.

Unfortunately this touches a lot of files. But it's probably the least
evil way to allow implementing code reuse ;)

Next up: hooking up feature pipeline JSON config using the load method.