Adds ability to run training entirely from a config file
Hooks up the same feature pipeline JSON config as used in alloys
to be used in the configuration file. This adds two example
configuration files -- I put them in predict, as otherwise
it couldn't find them if I put it in the train/resources folder
/shrug

To specify your own config file, you just use the `-c` or `--config`
parameter for the training app.