Adds HoldOutSetLogisticRegressionFurnace
This furnace randomly chooses data to split into two buckets,
one for training, the other for testing. The split is determined
by a parameter (80% for training is the default), and like
cross validation it builds a bunch of models and uses testing
data set to evaluate which is the best model.

TODO: Investigate how to extend this to pass in a real hold out set.