ENG-3135 Adds ValidationExamples to Alloy
This was not that trivial as I would have liked. The idea is that
post model training, we save a validation set based off of
the training docs that we save the results for. The idea is then
when we reload the jar, we can validate that the results
returned are the same -- thus being an early warning detection
system if underlying libraries change... Basically I made
Classification serializable, along with adding code to make it
easy to get a (document, predicition) saved on a per model basis.

At load time loading the examples is optional. If you do want to,
you need to provide the correct builder object and parameterization.

Unit tests will follow this commit.