ENG-3127: return Feature instances as significantFeatures
when documents are transformed into feature vectors at earlier
processing stages than the current model, it will not be possible
to return the "human-readable" feature representation with the
PredictResult

instead, switch to returning a Feature[_] type as the significant
features, so that the data returned upstream can be tailored to
what is available.

also, redefine TerminableTransformer#numDimensions as an Option[Int],
rather than an Int. I think it makes more sense for indeterminate
transformations (such as an unfrozen IndexTransformer) to return
None rather than returning an ever-changing value. this isn't
implemented in this commit (and if it wasn't such a PITA to split,
I'd make this a separate commit), but will feed into future cleanup
of this part of the system