ENG-3161 Passes through storing UUIDs and labels
Before we were only storing labels. This has been a TODO for a while
and is now complete. Basically we change to label UUIDs as being the
reference point. The annotations that come in, are now UUIDs rather
than a human readable label, so we just have to store the mapping
to be able to recover it, if we needed.

This change follows on from the idiml export data script, and ingests
a file that combines rules and label information. We change the BaseAlloy
then to store a mapping of UUID to an object "label". Loading the
single line JSON blob isn't that pretty, but this commit was trying
to be minimal.

We will probably have to change a few output scripts etc to make
sure we're outputting human versions of labels when we need to.