ENG-3162 Adds per label furnace configuration
We now have the ability to have per-label furnace configuration, meaning
we can setup XVal for one label, and SimpleLR for others... Basically this
was needed since after XVal, it's unlikely that all labels in the K-binary
case would have the same parameters. Thus if we were to train these models
again, we'd want to train them on what was found to be individually best,
rather than a global setting.