Adds ability to store training config in JarAlloy
I only wired this through for the JarAlloy, since I didn't want to
explicitly make the assumption that the BaseAlloy would need to know
about this. Perhaps it should. In anycase, we would now be storing
the entire training config with the jar. So knowing how it was
generated would be easy; haven't add the piece to get at it, since
it's not clear who or where that would happen.