ENG-3176 Add training blancer
This adds a DataSetScale trait, that then is implemented by two classes:
 - balanced binary
 - no-op
The idea is that the datagenerators will delegate to these classes. Rather
than this be something a datagenerator needs to do, I thought this way was
better since it allowed us to configure different strategies without having
to create brand new data generators.

Adds unit tests for them, as well as builder and factory methods so that
they will become accessible via configuration.