Adds metric and training summary classes
Oops, forgot to add these to the previous commit (or the preceeding one).

This adds a metrics class and metrics type and classes that encapsulate
metrics data, along with a training summary class which wraps a bunch
of metrics together. This will allow us to track all metrics that spark
exposes, plus a few of our own.
E.g. multi-class, multi-label, binary classifier, and hyperparemeter results. (+1 squashed commit)
Squashed commits:
[c61c6bf] Removes case from training summary