Add HasTrainingSummary trait & exposues getTrainingSummary...
This adds the HasTrainingSummary trait and exposes a getTrainingSummary
method in PredictModel. We need these methods to expose and save
a training summary, which is a collection of metrics. These metrics
will then be used to power things like model reporting, as well as
being able to historically packaging training performance with
an alloy.