Exposes task_type parameter in Trainers
We need to support passing through knowledge of whether a task is
to produce mutually exclusive labels or not. This will be useful
information for knowing how to deal with metrics, e.g. multiclass
vs multilabel.

The current distinctions for task type are:
`classification.single` for mutually exclusive tasks.
`classification.multiple` for non-mutually exclusive tasks.

This value is being passed through in the rule & label configuration
now.