Refactors MetricType to MetricTypes java enum
To get better metric & type safety, this introduces a java
enum instead of a scala enum, since in the java enum we can
add a property to the enum. In this case the property is the
metric class type, thus we can only house metrics in their
appopriate matching metric class type.

This then also includes all the necessary files to be touched
to achieve the refator. I left the metric trait in there just
incase we want to change the underlying assumptions easily.