Make Word2Vec transformer usable in a transformer pipelne
This PR will update the Word2Vec FeatureTransformer so
it can be used in the transformer pipeline. This includes
extending TerminableTransformer and adding in required methods.
Additionally, the parameter type passed in to Word2Vec's apply
will be changed to a Sequence of Features instead of Sequence of Strings,
so as to accept the results of Transformers that come before it.