Adds getSuggestedThresholds to alloy
This adds the threshold method to the alloy interface, and adds an implementation
to the BaseAlloy. Since the learning curve trainer and a few other places
would benefit from this, I decided to add this top level method.

Since it's not guaranteed that the underlying models will actually have
a suggested best F1 metric, I push it back to the caller to know
what to do if a particular label doesn't have a suggested threshold.