Refactors MetricsHelper a little and removes rest of hastrainingsummary
This adds the ability to pass through a metricClass to multi-label and
multi-class metrics methods. Since this will be required for the learning
curve trainer to be able to create binary label metrics while re-using
the metrics helper.

Also removes the last traces of hasTrainingSummary and PredictModels
making anonymous classes with that trait; they never needed to do that
in the first place.