Adds learning curve trainer, builder and tests
This change adds a learning curve trainer and everything
required to instantiate it.

The learning curve trainer runs under the assumption that it could
be dropped into the normal training task and everything would still work,
except that the alloy produced wouldn't be useful other than for
storing metrics.

The basic idea is that given a stream of training data, create logical
partitions over it (i.e. folds), that are split into a hold out set
and a bunch of training sets, where each training set is a 'portion' of
the total training set. Then for each fold, we traing an alloy for each
training set. We then evaluate all the alloys against the hold out set.
This then gives us a bunch of points for each fold, that we then average
across all the folds to produce a learning curve metric for F1, P & R.

The default folds are 5, since we want to test on 20% of the data, and
average across that. The portions come from the current system setup.
It's probably advised that the trainer should not be a xval one, since
that will take a while and has the potential to run into data sparsity issues...

Note: k-class models vs a single multinomial model should both work. However
for the multinomial case, we approximate it being a k-class model, so that
we can get individual per label learning curves... This might turn out to
yield non-sensical results, so it's possible the treatment of multi-nomial
models will need to change.

See https://idibon.atlassian.net/wiki/display/ENG/Generating+Learning+Curves+with+IDIML
for more details.