ENG-3172 Adds static methods to average metric subclasses
This adds utility methods to know how to average metric classes of
different metric types. This assumes we are passing in a homogeneous
sequence of metrics to average (well it asserts that that is the case).

I think this will be more useful and resusable than the crufty stuff
I am doing in LearningCurveTrainer. Unit tests to follow.