ENG-3172 Refactors fold generation
Moves the methods that are used to create a logical bunch of data
streams to be used as training and hold out sets for doing K-Fold
work, and puts them in their own trait and file. The reasoning
for this change is so that we can keep logical, reusuable functionality
together.

This also updates the unit tests to reflect this change.