Fixes bug in default thresholds in learning curves
I had flipped the defaults around. This is because in the
multi-lable case, we'll always have binary classifiers underpinning
things so a default of 0.5f makes sense there. In the mutually exclusive
label case, the classifier could be multinomial or k-binary in which
case take the lower threshold that makes sense, which is 1/#labels.