add simpler, more extensible Furnace interface
the Furnace interface is unfortunately pretty tightly coupled to
the workflow for generating models in Spark, and this limitation is
reflected in the configuration files: the feature pipeline(s) are
specified separately from the furnaces which are separate from the
data generation / sequencing step

and because this workflow doesn't match rules, we just hard-code
some logic for injecting rules models into the alloys as-needed
within the core trainer

this commit introduces a new trait hierarchy for Furnaces that will
extend to a much broader Alloy umbrella: classification models, span
models, rules, gangs, and topic models.

configuration of the Furnaces will follow the opaque hierarchical
approach used in the Feature Pipeline: essentially, every Furnace
implementation will define the configuration components that it
needs, and will be responsible for understanding its configuration
as-needed.