make Freezable a separate trait
with the addition of the SequenceGenerator and ChainPipeline graphs
for Span tasks, we may want to be able to freeze feature transforms
that are not TerminableTransformers.

so move TerminableTransform#freeze into a new trait, and change the
signature to allow implementations to return a frozen copy of the
original object, rather than Unit. this enables FeaturePipeline,
SequenceGenerator, and ChainPipeline to all implement Freezable,
rather than requiring those classes to define a similar-but-incompatible
method to accomplish the same