Adds Alloy Cross Validating & Competing Trainers/Classes
This adds a bunch of classes to facilitate alloy cross validation:
 - training an alloy & evaluating it.
 - averaging metrics from a bunch of results.
 - being able to put trainers head to head and choose
   the one that has the best F1 from the averaged results
   from cross validation.
 - config that can be used in the train task
 - unit tests

These classes mirror the class diagram on the RFC page.
This now gives the ability to get a sense for how a trained
model would work on "unseen" data.

These classes will allow a future refactor of the learning
curve implementation as well (in a future PR/Commit)