Use harmonic mean to average rules + ML predictions
Add a harmonic mean method to Classification which uses the
existing average method on each set of classifications
that came from the same source (ML vs. rules).

In order to keep track of the source of each PredictResult (and therefore
classification), the enum PredictTypeFlag is introduced. This flag has
values `model`, `rule`, `combined` to signify how a classification came to be.