Adds Alloy metric class
This adds an Alloy metric class to label "alloy level" evaluations.
This will make it easier to identify them downstream. This is a simpler
change that keeps metrics backwards compatible. The other option was
adding a field to training summaries to designate where the training
summary was created, if downstream consumption is too difficult that
might be the way to go...

Otherwise finishes adding toString methods on all metrics, as well
as pushing through the ability to give the metrics being averaged
a new metric class to facilitate using the alloy metric class.