ENG-2125: Use isRule flag in DocumentRules & Classification avg
Rename the Classification method `average` to `weighted_average`.
Add a new `average` method which returns the harmonic mean
of the `weighted_average`(s) of all isRule and non isRule Classifications.

Previously, the combined classification probability was calculated
as the average of all weighted probabilities (each probability
was weighted by the number of occurrences - this only made a difference
for rule occurrences as feature occurrences are hard coded to be 1).

The rationale to switching to harmonic mean between ML and rule
probability  was from a discovery where very high (though not forced)
analyst rules would make very little difference in the way idisage
calculated final probabilities. With the linear average
idiML was using, this was not the case but as the weighting with
occurrences would now skew combined probability too far in the
favor of rules, an average of averages (ML vs rules) was a better
representation of the combined proability.