Fix KClassKFP from failing on frozen pipelines
With the recent frozen change, this was missed because the test
was silently failing. Fixing this by instantiating a feature
pipeline for each label. So for K feature pipelines for M
labels we create K * M feature pipelines. This was because
we can't use "freeze" to clone a pipeline anymore. We can only
freeze a pipeline once.

This will probably have some memory impacts. /shrug