ENG-3329 Refactor LC
This refactors learning curves to use the alloy cross validation code.
So this simplifies a lot.

Otherwise this will now return a pre-packaged learning curve set of
metrics for consumption, along with the cross validation results
in separate training summaries. So it'll be up to the caller to
decide what to do.

TODO: finish off the unit tests.