ENG-3328 adds portion metric type
The metrics implementation is a bit messy. Rather than refactoring it
I'm doing the minimal to make it a bit easier to use... i.e.
this change is not strictly needed, but is the easiest change to make
that is backwards compatible with existing alloys out in the wild.

This now means that we need to inspect the sequence of metrics in
a summary for the portion... yes, less ideal, but the easiest
thing to do to not rely on training summary identifier to get
the portion information from...