ENG-3208 Add guards against small data in kfold data set creator
The k fold data set creator will now inspect the size of the incoming
data set. If it less than 4 * the number of folds, it'll just create
a single fold from that data, with the evaluation and training being
the same set. I thought that this would be a safer and cleaner option
than trying to inspect the data items themselves.